% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{col_meta_GSE70138}
\alias{col_meta_GSE70138}
\title{#TODO}
\format{
An object of class \code{data.frame} with 118050 rows and 8 columns.
}
\usage{
col_meta_GSE70138
}
\description{
#TODO
}
\keyword{datasets}
