% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{col_meta_GSE92742}
\alias{col_meta_GSE92742}
\title{#TODO}
\format{
An object of class \code{data.frame} with 473647 rows and 12 columns.
}
\usage{
col_meta_GSE92742
}
\description{
#TODO
}
\keyword{datasets}
