% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugScore.R
\name{DrugScore}
\alias{DrugScore}
\title{Calculate drug score}
\usage{
DrugScore(
  cell_metadata,
  cluster_degs,
  cluster_drugs,
  tissue,
  gse70138_gctx_path,
  gse92742_gctx_path,
  clusters = NULL,
  case = NULL,
  fda_drugs_only = TRUE
)
}
\arguments{
\item{cell_metadata}{A data.frame of cell metadata. It must have a column 
named 'cluster' indicating which cluster cells belong, and a column named 
'sample' indicating which sample cells belong.}

\item{cluster_degs}{A list of differential gene expression profiles for 
each cluster.}

\item{cluster_drugs}{Drug repurposing result from GetDrug function.}

\item{tissue}{Reference tissue. If one used 'lung_rankMatrix.txt' in 
GetDrugRef function, then the Reference tissue is lung. Please use " " 
instead of "-" in tissue name. For example, while 
'haematopoietic-and-lymphoid-tissue' is the prefix of the drug reference 
files, the corresponding tissue name is "haematopoietic and lymphoid tissue".}

\item{gse70138_gctx_path}{The gctx file contains drug responses from GSE70138 
dataset (https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE70138).}

\item{gse92742_gctx_path}{The gctx file contains drug responses from GSE92742 
dataset (https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE92742)..}

\item{clusters}{Select which clusters (cell types) to be used for drug score 
estimation. By default, it uses all clusters.}

\item{case}{A vector containing case sample names.}

\item{fda_drugs_only}{logical; if TRUE, will only return FDA-approved drugs, 
else, will return all drugs/compounds.}
}
\value{
A data frame of drug score, P-value and FDR.
}
\description{
The drug score is a comprehensive estimation of drug therapeutic 
effects using all or a selected set of clusters.
}
\details{
This function calculates drug score using cellular proportion of 
clusters, the significance of reversal in DEGs' expressions, and the ratio of 
the reversed genes.
}
