% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDrug.R
\name{GetDrug}
\alias{GetDrug}
\title{Mono-drug Repurposing.}
\usage{
GetDrug(
  gene.data = NULL,
  drug.ref.profiles = NULL,
  repurposing.unit = "drug",
  CEG.threshold = 0.05,
  connectivity = "negative",
  drug.type = "FDA"
)
}
\arguments{
\item{drug.ref.profiles}{A list contains tissue specific drug reference Profiles from GetDrugRef function.}

\item{repurposing.unit}{The parameter of either "treatment" or "drug", which indicates if user want the function to test drug repurposing p value at treatment level or drug level. The default is "treatment", which treats the drug data from different cell lines separately.}

\item{CEG.threshold}{The p value threshold to select the consistently differential expressed genes (CEGs). The default value is 0.05.}

\item{connectivity}{The type of connectivity, either "negative" or "positive". Negative connectivity is used when the query data is the differential scores from disease data, and it will repurpose drugs that can potentially reverse the query disease phenotype. Positive connectivity is used when the query data is from a drug profile, and it will return the drugs that are similar to the query drug. The default value is "negative".}

\item{drug.type}{The parameter of either "FDA" or "compounds" or "all", which indicates if user want the function to identify FDA-approved drugs or compounds or both, respectively.The default value is "FDA".}
}
\value{
A list of mono-drugs for every cell type.
}
\description{
It identify mono-drug therapy for every cell type.
}
\details{
This function allows user to use the differential expression data of every case cell type to query against reference drug response profiles.This function is a reverised version of drug.identification from DrInsight package.
}
