% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDrugRef.R
\name{GetDrugRef}
\alias{GetDrugRef}
\title{Load and Process Drug Reference Profiles.}
\usage{
GetDrugRef(drug.response.path = NULL, probe.to.genes = NULL, drug.info = NULL)
}
\arguments{
\item{drug.response.path}{The local path and the name of the tissue specific drug rank matrix.}

\item{probe.to.genes}{A data.frame contains gene IDs (the IDs used in drug rank matrix) and official gene symbol. This files was automately generated with drug rank matrix.}

\item{drug.info}{A data.frame contains drug information. This file was automately generated with drug rank matrix.}
}
\description{
This function allows user to load in the tissue specific drug rank matrix.
}
\details{
This function is a reverised version of get.cmap.ref from DrInsight package. The tissue specific drug rank matrix is tranformed from L1000data (GEO: GSE92742 and GSE70138) using PrepareReference function.
}
