% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGene.R
\name{GetGene}
\alias{GetGene}
\title{Get Differntially Expressed Genes.}
\usage{
GetGene(
  SC.integrated = SC.integrated,
  Case = Case,
  Control = Control,
  min.cells = 3
)
}
\arguments{
\item{SC.integrated}{A Seurat object of aligned cells from SCalignment function.}

\item{Case}{A vector contains names of case samples.}

\item{Control}{A vector contains names of control samples.}

\item{min.cells}{The minimum number of cells for a cell type. A cell type is omitted if it has less cells than the minimum number.}
}
\value{
A list of differnential gene expression profiles for every cell type.
}
\description{
Identify differnential gene expression profiles for case cells compared to control cells.
}
\details{
This function identifies differentially expressed genes between paired case and control cells for every type of cell. If there is no cell type annotation, it uses every cluster as a cell type.
}
