% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{load_luekemia}
\alias{load_luekemia}
\title{Load Luekemia dataset analyzed in the Asgard publication}
\usage{
load_luekemia(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{The directory to save downloaded datasets. If not giving,
the datasets are saved under the default cache directory of `BiocFileCache` 
library.}
}
\value{
A list of Seurat objects.
}
\description{
This function load Luekimia datasets studied in the paper. At the first 
run, the function will downlaod necessary datasets from GSE and save them 
under a cache directory using `BiocFileCache`` library.
}
