% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareReference.R
\name{PrepareReference}
\alias{PrepareReference}
\title{Prepare Drug Reference.}
\usage{
PrepareReference(
  cell.info = NULL,
  gene.info = NULL,
  GSE70138.sig.info = NULL,
  GSE92742.sig.info = NULL,
  GSE70138.gctx = NULL,
  GSE92742.gctx = NULL,
  Output.Dir = "./"
)
}
\arguments{
\item{cell.info}{The local path and the name of the cell.info text file. It's downloaded from https://ftp.ncbi.nlm.nih.gov/geo/series/GSE70nnn/GSE70138/suppl/GSE70138_Broad_LINCS_cell_info_2017-04-28.txt.gz .}

\item{gene.info}{The local path and the name of the gene.info text file. It's downloaded from https://ftp.ncbi.nlm.nih.gov/geo/series/GSE70nnn/GSE70138/suppl/GSE70138_Broad_LINCS_gene_info_2017-03-06.txt.gz .}

\item{GSE70138.sig.info}{The local path and the name of the cell.info text file. It's downloaded from https://ftp.ncbi.nlm.nih.gov/geo/series/GSE70nnn/GSE70138/suppl/GSE70138_Broad_LINCS_sig_info_2017-03-06.txt.gz .}

\item{GSE92742.sig.info}{The local path and the name of the cell.info text file. It's downloaded from https://ftp.ncbi.nlm.nih.gov/geo/series/GSE92nnn/GSE92742/suppl/GSE92742_Broad_LINCS_sig_info.txt.gz .}

\item{GSE70138.gctx}{The local path and the name of the cell.info text file. It's downloaded from https://ftp.ncbi.nlm.nih.gov/geo/series/GSE70nnn/GSE70138/suppl/GSE70138_Broad_LINCS_Level5_COMPZ_n118050x12328_2017-03-06.gctx.gz .}

\item{GSE92742.gctx}{The local path and the name of the cell.info text file. It's downloaded from https://ftp.ncbi.nlm.nih.gov/geo/series/GSE92nnn/GSE92742/suppl/GSE92742_Broad_LINCS_Level5_COMPZ.MODZ_n473647x12328.gctx.gz .}

\item{Output.Dir}{The output directory for the generated files.}
}
\description{
Prepare tissue specific drug reference Profiles from L1000 drug response data.
}
\details{
This function converts L1000 data to the tissue specific drug rank matrix.
}
