% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepRefv2.R
\name{PrepRefv2}
\alias{PrepRefv2}
\title{Prepare Drug Reference using a single updated dataset.}
\usage{
PrepRefv2(
  cell.info = NULL,
  gene.info = NULL,
  sig.info = NULL,
  gctx = NULL,
  Output.Dir = "./"
)
}
\arguments{
\item{cell.info}{The local path and name of the cell.info text file (from LINCS L1000).}

\item{gene.info}{The local path and name of the gene.info text file.}

\item{sig.info}{The local path and name of the signature info (sig.info) text file for the updated dataset.}

\item{gctx}{The local path and name of the GCTX file for the updated dataset.}

\item{Output.Dir}{The output directory for the generated files.}
}
\description{
Prepare tissue-specific drug reference profiles from a single L1000 drug response dataset.
}
\details{
This function converts L1000 data to the tissue specific drug rank matrix for a single dataset.
}
