% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCplasticity.R
\name{SCplasticity}
\alias{SCplasticity}
\title{Sinlge-cell Plasticity.}
\usage{
SCplasticity(SC.integrated = SC.data, Case = NULL)
}
\arguments{
\item{SC.integrated}{A Seurat object of aligned single cells from SCalignment function.}

\item{Case}{A vector contains names of case samples.}
}
\value{
A data frame of plasticity, normailized plasticity and cell type coverage.
}
\description{
It determines the plasticity of each cell type.
}
\details{
This function estimate the entropy of every cell in the case samples. For each cell type, it use the median entropy value as the plasticity of each cell type.
}
