% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopCombination.R
\name{TopCombination}
\alias{TopCombination}
\title{Combination Drug Selection.}
\usage{
TopCombination(
  Drug.combination = Drug.combinations,
  Combination.FDR = 0.1,
  Min.combination.score = 1
)
}
\arguments{
\item{Drug.combination}{raw drug combination result from DrugCombination function.}

\item{Combination.FDR}{The FDR threshold to select drug combination. The default value is 0.1.}

\item{Min.combination.score}{The Combination therapeutic score threshold to select drug combination. The default value is 1.}
}
\value{
A data frame of selected drug combinations.
}
\description{
Select drug combinations by combination therapeutic score and FDR of combination therapeutic score.
}
\details{
Input raw drug combination result and return the top drug combinations.
}
