% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopDrug.R
\name{TopDrug}
\alias{TopDrug}
\title{Single Drug Selection.}
\usage{
TopDrug(
  SC.integrated = SC.data,
  Drug.data = Drug.ident.res,
  Drug.FDR = 0.1,
  FDA.drug.only = TRUE,
  Case = NULL
)
}
\arguments{
\item{SC.integrated}{Single-cell alignment result from Seurat.}

\item{Drug.data}{Drug repurosing result from drug.ident function.}

\item{Drug.FDR}{The FDR threshold to select drug. The default value is 0.1.}

\item{FDA.drug.only}{logical; if TRUE, will only return FDA-approved drugs.}

\item{Case}{An vector of case (diseased) samples.Only case sammples are involved in the calculation of coverage.}
}
\value{
A data frame of selected drugs with summary of cell coverage.
}
\description{
Select single drugs for every cell population by FDR and drug type, and summarize cell coverage for selected drugs.
}
\details{
Input raw drug repurosing result and return the top drugs with summary of cell coverage.
}
