% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugScore.R
\name{DrugScore}
\alias{DrugScore}
\title{Treatment Efficacy of the Drug.}
\usage{
DrugScore(
  SC.integrated = SC.data,
  Gene.data = Gene.list,
  Cell.type = NULL,
  Drug.data = Drug.ident.res,
  FDA.drug.only = TRUE,
  GSE92742.gctx = NULL,
  GSE70138.gctx = NULL,
  Case = NULL,
  Tissue = "breast"
)
}
\arguments{
\item{SC.integrated}{A Seurat object of aligned single cells from Seurat.}

\item{Gene.data}{A list of differnential gene expression profiles for every cell type. It's from GetGene function.}

\item{Drug.data}{A list of mono-drugs for every cell type. It's from GetDrug function.}

\item{FDA.drug.only}{logical; if TRUE, will only return FDA-approved drugs.}

\item{GSE92742.gctx}{The local path and the name of the gctx file from GSE92742 dataset (https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE92742).}

\item{GSE70138.gctx}{The local path and the name of the gctx file from GSE70138 dataset (https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE70138).}

\item{Case}{A vector contains names of case samples.}

\item{Tissue}{Reference tissue. Usually it's same with the drug reference tissue.}

\item{Drug.FDR}{The FDR threshold to select drug. The default value is 0.1.}
}
\value{
A data frame of drug combinations with therapeutics scores.
}
\description{
It evaluates treatment efficacy to identify drug that can best reverse the target genes’ expression in diseased cells in case samples.
}
\details{
This function evaluates treatment efficacy and ranks drugs using therapeutics score, which integrates gene responses to multiple drugs, the proportion of genes, and cells treated by drugs.
}
