% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugScore.R
\name{DrugScore}
\alias{DrugScore}
\title{Drug score}
\usage{
DrugScore(
  SC.integrated = SC.data,
  Gene.data = Gene.list,
  Cell.type = NULL,
  Drug.data = Drug.ident.res,
  FDA.drug.only = TRUE,
  GSE92742.gctx = NULL,
  GSE70138.gctx = NULL,
  Case = NULL,
  Tissue = "breast"
)
}
\arguments{
\item{SC.integrated}{A Seurat object of aligned single cells from Seurat.}

\item{Gene.data}{A list of differnential gene expression profiles for every cell type. It's from GetGene function.}

\item{Cell.type}{Select which clusters (cell types) to be used for drug score estimation. By default, it uses all clusters.}

\item{Drug.data}{A list of mono-drugs for every cluster. It's from GetDrug function.}

\item{FDA.drug.only}{logical; if TRUE, will only return FDA-approved drugs, else, will return all inputted drugs/compounds}

\item{GSE92742.gctx}{The gctx file contains drug responses from GSE92742 dataset (https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE92742).}

\item{GSE70138.gctx}{The gctx file contains drug responses from GSE70138 dataset (https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE70138).}

\item{Case}{A vector contains case sample names.}

\item{Tissue}{Reference tissue. If one used lung_rankMatrix.txt in GetDrugRef function, then the Reference tissue is lung.}
}
\value{
A data frame of drug score, P-value and FDR.
}
\description{
The drug score is a comprehensive estimation of drug therapeutic effects acrossing all or selected single cell clusters.
}
\details{
This function estimates drug treatment efficacy using therapeutics score, fisher's combined P-value and FDR.
}
